feat(EditWork): added google places api for work
