+++ src/components/UserProfile.js
@@ -39,6 +39,7 @@ class UserProfile extends React.Component {
         <h6>Slack: {user.slack}</h6>
         <h6>Github: {user.github}</h6>
         <h6>Linkedin: {user.linkedin}</h6>
+        <h6>Works at: {user.workInfo}</h6>
         {user.projects ?
           <h5>Projects: {user.projects.length && this.renderProjects(user.projects)}</h5>
           : null
